pyproject.toml
[build-system]
requires = ["setuptools>=64", "wheel"]
build-backend = "setuptools.build_meta"

[project]
name = "genmol"
readme = "README.md"
description = "GenMol A Drug Discovery Generalist with Discrete Diffusion"
authors = [{ name = "NVIDIA", email = "bionemofeedback@nvidia.com" }]
requires-python = ">=3.10"
license = { file = "LICENSE/license_code.txt" }
dynamic = ["version"]
dependencies = [
    # bionemo sub-packages
    'bionemo-moco==0.0.2.1',
    # external
    'torch==2.6.0',
    'numpy==1.26.4',
    'jaxtyping>=0.2.34',
    'pot>=0.9.5',
    'scikit-learn>=1.2.2',
    'matplotlib>=3.3.2',
    'einops==0.7.0',
    'wandb==0.13.5',
    'pandas==2.1.0',
    'datasets==2.18.0',
    'hydra-core==1.3.2',
    'lightning==2.5.1',
    'omegaconf==2.3.0',
    'transformers==4.52.4',
    'safe-mol==0.1.13',
    'pytdc==0.4.1',
    'easydict==1.13',
    'openbabel-wheel==3.1.1.22'
]

[tool.setuptools.packages.find]
where = ["src"]
include = ["*"]
namespaces = true


[tool.setuptools.dynamic]
version = { file = "VERSION" }
